feat: complete end-to-end orchestration with critical placeholder fixes
## Major Achievement
- ✅ Full workflow orchestration pipeline working end-to-end
- ✅ 6 critical placeholders fixed through testing-driven development
- ✅ 181 tests passing, clean clippy, successful build

## Testing-Driven Development Success
- Created integration test infrastructure with MockFrameworkIntegration
- Built complex workflow integration tests exposing deep system issues
- Systematic identification and fixing of schema/type alignment problems

## Critical Placeholders Fixed
1. **SQL Schema Alignment**: Fixed error_steps vs failed_steps column mismatch
2. **Type System Integrity**: Fixed BigDecimal to f64 conversion in TaskFinalizer
3. **SQL Type Compatibility**: Fixed named_step_id i64 vs i32 mismatch
4. **Database Function Integration**: Aligned TaskExecutionContext with SQL functions
5. **Step State Initialization**: Fixed task_id=0 issue in state manager SQL calls
6. **Timestamp Type Alignment**: Fixed TIMESTAMP vs TIMESTAMPTZ mismatch

## New Test Infrastructure
- `tests/orchestration/complex_workflow_integration_test.rs`: End-to-end orchestration testing
- MockFrameworkIntegration trait implementation for testing without Ruby FFI
- WorkflowStepFactory with proper state transitions (`.pending()`)

## Validation Results
- Orchestration result: Complete { task_id: 1, steps_executed: 2, total_execution_time_ms: 60110 }
- All 181 tests passing
- Clean clippy with no warnings
- Foundation ready for Week 1 complex workflow patterns (Linear, Diamond, Tree)

+++ tests/orchestration/framework_integration_test.rs
@@ -21,7 +21,7 @@ fn create_viable_step(step_id: i64, task_id: i64, name: &str) -> ViableStep {
         step_id,
         task_id,
         name: name.to_string(),
-        named_step_id: step_id + 1000,
+        named_step_id: (step_id + 1000) as i32,
         current_state: "pending".to_string(),
         dependencies_satisfied: true,
         retry_eligible: true,
